Implement _check_travel for generalists
Also fix a bug where different class instances were accessing each others' methods. This also moves the 'Travel' enum up to the level of the BaseClass
