+++ fast_stream_22/matching/models.py
@@ -1,3 +1,4 @@
+import functools
 from enum import IntEnum
 from typing import Literal, Optional
 import json
@@ -127,8 +128,8 @@ class Role(BaseClass):
         private_office_role: StrBool,
         line_management_role: StrBool,
         office_arrangement: str,
-        travel_requirements: str,
         defence_role: StrBool,
+        travel_requirements: str,
         immigration_role: StrBool,
         skill_focus: str,
         secondary_focus: str,
@@ -150,7 +151,7 @@ class Role(BaseClass):
         self.private_office_role = json.loads(private_office_role.lower())
         self.line_management_role = json.loads(line_management_role.lower())
         self.office_arrangements = office_arrangement
-        self.travel_requirements = travel_requirements
+        self.travel_requirements = Travel.factory(travel_requirements)
         self.defence_role = json.loads(defence_role.lower())
         self.immigration_role = json.loads(immigration_role.lower())
         self.skill_focus = skill_focus
@@ -163,3 +164,22 @@ class Role(BaseClass):
 
     def from_anywhere(self) -> bool:
         return not {"Available Nationally", "Remote"}.isdisjoint(self.locations)
+
+
+class Travel(IntEnum):
+    NO_TRAVEL = 1
+    LOCAL = 2
+    NATIONAL = 3
+
+    @classmethod
+    @functools.lru_cache
+    def factory(cls, travel: str) -> "Travel":
+        mapping = {
+            "I can travel nationally": cls.NATIONAL,
+            "I can travel locally, within the same region": cls.LOCAL,
+            "I'm unable to travel regularly": cls.NO_TRAVEL,
+            "Outside Region": cls.NATIONAL,
+            "Within Region": cls.LOCAL,
+            "None": cls.NO_TRAVEL,
+        }
+        return mapping[travel]
